Moved output time zone options into tool options
