+++ tests/cli/psteal_tool.py
@@ -2,7 +2,6 @@
 # -*- coding: utf-8 -*-
 """Tests for the psteal CLI tool."""
 
-import argparse
 import io
 import os
 import unittest
@@ -22,25 +21,6 @@ class PstealToolTest(test_lib.CLIToolTestCase):
   # pylint: disable=protected-access
 
   _BDE_PASSWORD = 'bde-TEST'
-
-  _EXPECTED_OUTPUT_TIME_ZONE_OPTION = """\
-usage: psteal_test.py [--output_time_zone TIME_ZONE]
-
-Test argument parser.
-
-optional arguments:
-  --output_time_zone TIME_ZONE, --output-time-zone TIME_ZONE
-                        time zone of date and time values written to the
-                        output, if supported by the output format. Output
-                        formats that support this are: dynamic and l2t_csv.
-                        Use "list" to see a list of available time zones.
-"""
-
-  _EXPECTED_PROCESSING_OPTIONS = """\
-usage: psteal_test.py
-
-Test argument parser.
-"""
 
   _STORAGE_FILENAME_TEMPLATE = r'\d{{8}}T\d{{6}}-{filename}.plaso'
 
@@ -92,35 +72,6 @@ Test argument parser.
     expected_storage_filename = self._STORAGE_FILENAME_TEMPLATE.format(
         filename='bar')
     self.assertRegex(storage_filename, expected_storage_filename)
-
-  def testParseOutputTimeZoneOption(self):
-    """Tests the _ParseOutputTimeZoneOption function."""
-    test_tool = psteal_tool.PstealTool()
-
-    options = test_lib.TestOptions()
-
-    test_tool._ParseOutputTimeZoneOption(options)
-    self.assertIsNone(test_tool._output_time_zone)
-
-    options.output_time_zone = 'list'
-    test_tool._ParseOutputTimeZoneOption(options)
-    self.assertIsNone(test_tool._output_time_zone)
-
-    options.output_time_zone = 'CET'
-    test_tool._ParseOutputTimeZoneOption(options)
-    self.assertEqual(test_tool._output_time_zone, 'CET')
-
-  def testAddOutputTimeZoneOption(self):
-    """Tests the AddOutputTimeZoneOption function."""
-    argument_parser = argparse.ArgumentParser(
-        prog='psteal_test.py', description='Test argument parser.',
-        add_help=False, formatter_class=test_lib.SortedArgumentsHelpFormatter)
-
-    test_tool = psteal_tool.PstealTool()
-    test_tool.AddOutputTimeZoneOption(argument_parser)
-
-    output = self._RunArgparseFormatHelp(argument_parser)
-    self.assertEqual(output, self._EXPECTED_OUTPUT_TIME_ZONE_OPTION)
 
   def testFailWhenOutputAlreadyExists(self):
     """Test to make sure the tool raises when the output file already exists."""
